+++ apps/package-manager/packages/generic/src/generateExpectations/nrk/expectations.ts
@@ -20,6 +20,7 @@ import {
 	generateQuantelClipPreview,
 	generateJsonDataCopy,
 	generatePackageCopyFileProxy,
+	generatePackageLoudness,
 } from './expectations-lib'
 import { getSmartbullExpectedPackages, shouldBeIgnored } from './smartbull'
 import { TEMPORARY_STORAGE_ID } from './lib'
@@ -259,6 +260,15 @@ function getSideEffectOfExpectation(
 				expectations[preview.id] = preview
 			}
 		}
+
+		if (expectation0.sideEffect?.loudnessPackageSettings) {
+			const loudness = generatePackageLoudness(
+				expectation,
+				expectation0.sideEffect?.loudnessPackageSettings,
+				settings
+			)
+			expectations[loudness.id] = loudness
+		}
 	} else if (expectation0.type === Expectation.Type.QUANTEL_CLIP_COPY) {
 		const expectation = expectation0 as Expectation.QuantelClipCopy
 
@@ -302,6 +312,15 @@ function getSideEffectOfExpectation(
 				)
 				expectations[preview.id] = preview
 			}
+		}
+
+		if (expectation0.sideEffect?.loudnessPackageSettings) {
+			const loudness = generatePackageLoudness(
+				expectation,
+				expectation0.sideEffect?.loudnessPackageSettings,
+				settings
+			)
+			expectations[loudness.id] = loudness
 		}
 	}
 	return expectations