+++ apps/package-manager/packages/generic/src/generateExpectations/nrk/types.ts
@@ -61,4 +61,5 @@ export enum PriorityAdditions {
 	THUMBNAIL = 1002,
 	PREVIEW = 1003,
 	DEEP_SCAN = 1004,
+	LOUDNESS_SCAN = 1010,
 }