+++ shared/packages/expectationManager/src/evaluationRunner/evaluateExpectationStates/waiting.ts
@@ -45,7 +45,14 @@ export async function evaluateExpectationStateWaiting({
 				)
 
 				const newStatus: Partial<TrackedExpectation['status']> = {}
-				if (readyToStart.sourceExists !== undefined) newStatus.sourceExists = readyToStart.sourceExists
+				{
+					const sourceExists = readyToStart.ready || readyToStart.sourceExists
+					if (sourceExists !== undefined) newStatus.sourceExists = sourceExists
+				}
+				{
+					const isPlaceholder = readyToStart.ready ? false : readyToStart.isPlaceholder
+					if (isPlaceholder !== undefined) newStatus.sourceIsPlaceholder = isPlaceholder
+				}
 
 				if (readyToStart.ready) {
 					tracker.trackedExpectationAPI.updateTrackedExpectationStatus(trackedExp, {