+++ shared/packages/worker/src/worker/accessorHandlers/localFolder.ts
@@ -4,8 +4,13 @@ import fs from 'fs'
 import {
 	PackageReadInfo,
 	PutPackageHandler,
-	AccessorHandlerResult,
 	SetupPackageContainerMonitorsResult,
+	AccessorHandlerRunCronJobResult,
+	AccessorHandlerCheckHandleReadResult,
+	AccessorHandlerCheckHandleWriteResult,
+	AccessorHandlerCheckPackageContainerWriteAccessResult,
+	AccessorHandlerCheckPackageReadAccessResult,
+	AccessorHandlerTryPackageReadResult,
 } from './genericHandle'
 import {
 	Accessor,
@@ -20,6 +25,8 @@ import { GenericWorker } from '../worker'
 import { GenericFileAccessorHandle, LocalFolderAccessorHandleType } from './lib/FileHandler'
 import { MonitorInProgress } from '../lib/monitorInProgress'
 import { compareResourceIds } from '../workers/windowsWorker/lib/lib'
+import { defaultCheckHandleRead, defaultCheckHandleWrite } from './lib/lib'
+import { mkdirp } from 'mkdirp'
 
 const fsStat = promisify(fs.stat)
 const fsAccess = promisify(fs.access)
@@ -71,19 +78,17 @@ export class LocalFolderAccessorHandle<Metadata> extends GenericFileAccessorHand
 		return path.join(this.folderPath, this.filePath)
 	}
 
-	checkHandleRead(): AccessorHandlerResult {
-		if (!this.accessor.allowRead) {
-			return { success: false, reason: { user: `Not allowed to read`, tech: `Not allowed to read` } }
-		}
+	checkHandleRead(): AccessorHandlerCheckHandleReadResult {
+		const defaultResult = defaultCheckHandleRead(this.accessor)
+		if (defaultResult) return defaultResult
 		return this.checkAccessor()
 	}
-	checkHandleWrite(): AccessorHandlerResult {
-		if (!this.accessor.allowWrite) {
-			return { success: false, reason: { user: `Not allowed to write`, tech: `Not allowed to write` } }
-		}
+	checkHandleWrite(): AccessorHandlerCheckHandleWriteResult {
+		const defaultResult = defaultCheckHandleWrite(this.accessor)
+		if (defaultResult) return defaultResult
 		return this.checkAccessor()
 	}
-	private checkAccessor(): AccessorHandlerResult {
+	private checkAccessor(): AccessorHandlerCheckHandleWriteResult {
 		if (this.accessor.type !== Accessor.AccessType.LOCAL_FOLDER) {
 			return {
 				success: false,
@@ -101,7 +106,7 @@ export class LocalFolderAccessorHandle<Metadata> extends GenericFileAccessorHand
 		}
 		return { success: true }
 	}
-	async checkPackageReadAccess(): Promise<AccessorHandlerResult> {
+	async checkPackageReadAccess(): Promise<AccessorHandlerCheckPackageReadAccessResult> {
 		try {
 			await fsAccess(this.fullPath, fs.constants.R_OK)
 			// The file exists and can be read
@@ -117,7 +122,7 @@ export class LocalFolderAccessorHandle<Metadata> extends GenericFileAccessorHand
 		}
 		return { success: true }
 	}
-	async tryPackageRead(): Promise<AccessorHandlerResult> {
+	async tryPackageRead(): Promise<AccessorHandlerTryPackageReadResult> {
 		try {
 			// Check if we can open the file for reading:
 			const fd = await fsOpen(this.fullPath, 'r')
@@ -128,20 +133,26 @@ export class LocalFolderAccessorHandle<Metadata> extends GenericFileAccessorHand
 			if (err && (err as any).code === 'EBUSY') {
 				return {
 					success: false,
+					packageExists: true,
 					reason: { user: `Not able to read file (file is busy)`, tech: `${stringifyError(err, true)}` },
 				}
 			} else if (err && (err as any).code === 'ENOENT') {
-				return { success: false, reason: { user: `File does not exist`, tech: `${stringifyError(err, true)}` } }
+				return {
+					success: false,
+					packageExists: false,
+					reason: { user: `File does not exist`, tech: `${stringifyError(err, true)}` },
+				}
 			} else {
 				return {
 					success: false,
+					packageExists: false,
 					reason: { user: `Not able to read file`, tech: `${stringifyError(err, true)}` },
 				}
 			}
 		}
 		return { success: true }
 	}
-	async checkPackageContainerWriteAccess(): Promise<AccessorHandlerResult> {
+	async checkPackageContainerWriteAccess(): Promise<AccessorHandlerCheckPackageContainerWriteAccessResult> {
 		try {
 			await fsAccess(this.folderPath, fs.constants.W_OK)
 			// The file exists
@@ -188,6 +199,8 @@ export class LocalFolderAccessorHandle<Metadata> extends GenericFileAccessorHand
 		await this.clearPackageRemoval(this.filePath)
 
 		const fullPath = this.workOptions.useTemporaryFilePath ? this.temporaryFilePath : this.fullPath
+
+		await mkdirp(path.dirname(fullPath)) // Create folder if it doesn't exist
 
 		// Remove the file if it exists:
 		let exists = false
@@ -249,7 +262,7 @@ export class LocalFolderAccessorHandle<Metadata> extends GenericFileAccessorHand
 	async removeMetadata(): Promise<void> {
 		await this.unlinkIfExists(this.metadataPath)
 	}
-	async runCronJob(packageContainerExp: PackageContainerExpectation): Promise<AccessorHandlerResult> {
+	async runCronJob(packageContainerExp: PackageContainerExpectation): Promise<AccessorHandlerRunCronJobResult> {
 		// Always check read/write access first:
 		const checkRead = await this.checkPackageContainerReadAccess()
 		if (!checkRead.success) return checkRead
@@ -327,7 +340,7 @@ export class LocalFolderAccessorHandle<Metadata> extends GenericFileAccessorHand
 		return this.fullPath + '_metadata.json'
 	}
 
-	private async checkPackageContainerReadAccess(): Promise<AccessorHandlerResult> {
+	private async checkPackageContainerReadAccess(): Promise<AccessorHandlerRunCronJobResult> {
 		try {
 			await fsAccess(this.folderPath, fs.constants.R_OK)
 			// The file exists